refactor: move tests into functional and skip tests on windows
