+++ test/torchaudio_unittest/functional/autograd_cuda_test.py
@@ -1,8 +1,11 @@
 import torch
-from .lfilter_autograd_impl import Autograd
+from .autograd_impl import Autograd
 from torchaudio_unittest import common_utils
+import platform
+import unittest
 
 
+@unittest.skipIf(platform.system() == 'Windows', "not supported on windows")
 @common_utils.skipIfNoCuda
 class TestAutogradLfilterCUDA(Autograd, common_utils.PytorchTestCase):
     dtype = torch.float64
